refactor(#615): carve kanban.rs into kanban/{mod,state,format}.rs -- pure move
Line-coverage-asserted mechanical split (the proven carve mechanism, 4th
application): the generator segments the original 2,090-line file into 21
contiguous segments and asserts their concatenation reproduces the source
byte-for-byte, so every line is claimed exactly once. No logic changes.

- kanban/state.rs (387): CardStatus, Priority, Action, transition + the
  state-machine tests.
- kanban/format.rs (578): list/view/JSON/goal-banner rendering,
  CardSummary, priority_tag + the formatting tests.
- kanban/mod.rs (1,162): Card, AgentWorkload, map_card_row, CRUD wrappers,
  update_card/resolve_labels + DB-integration tests; re-exports state::*
  and format::* so every crate::kanban:: path is unchanged.

37 declared scaffold lines (module headers, mod/pub use decls, use decls,
test-module wrappers); cargo fmt deferred to its own commit so the moved
lines stay verbatim-diffable against the original.
